SoundEventDetection/requirements.txt
soundfile
librosa
tqdm
numpy
pandas
pypeln
h5py
torch
scikit-learn
tabulate
scipy
PyYAML
sed_eval
scikit-multilearn
